Keep return values from yield alive while mapping
Preivously, it was possible for the GC to free values that were
returned in earlier iterations during mapping. Since the GC does
not know how to traverse the Rust heap, the only thing reachable
by it is what is on the call stack. After an iteration of map
finishes, Rust puts the VALUE from it into the Rust heap, making
it a target for collection in futher iterations.

The fix is to have a Ruby array keeping track of values in all
map iterations on the call stack at all times while mapping.
The GC is able to recognize the Ruby arrry and refrain from
collecting the elements.