Use marshal_dump/load to implement marshaling instead
The main reason for this change is that special attribute
handles which come from a set had their parent instance
varaiable marhsalized. Also, _dump_data and _load_data are
undocumented.

The instance variables are only their to keep the parent
alive in case direct reference to the set is lost elsewhere,
so they should not show up in the marhsaled data.