Keep return value from #dup alive when deep_duping AttributeSet
When we are looping over the attribute set and constrcuting the
new set, only one VALUE is on top of the stack at any given time.
VALUEs from the previous iteration are put into Rust's heap,
which Ruby's GC does not understand how to traverse.

It may happen that the #dup Ruby call on the cached value triggers
a GC run. That run could end up freeing the results from #dup from
previous iterations, since those results are on Rust's heap and are
thus unreachable.

The fix is to put a Ruby array on the call stack and put every #dup
return value in it. The GC understands how to traverse an Ruby array
and refrains from collecting those values in result.